Move schedule parser into new scheduling path
